+++ tests/test_checks.py
@@ -1,7 +1,9 @@
 import pytest
 
+from spinta import commands
 from spinta.testing.manifest import load_manifest_and_context
 from spinta.manifests.tabular.helpers import TabularManifestError
+from spinta.exceptions import InvalidValue
 
 
 def test_enum_level(tmpdir, rc):
@@ -18,3 +20,52 @@ def test_enum_level(tmpdir, rc):
     assert str(e.value) == (
         "None:6: Enum's do not have a level, but level '3' is given."
     )
+
+
+def test_enum_type_integer(tmpdir, rc):
+    context, manifest = load_manifest_and_context(rc, '''
+    d | r | b | m | property | type    | prepare
+    datasets/gov/example     |         |
+                             |         |
+      |   |   | Data         |         |
+      |   |   |   | value    | integer |
+                             | enum    | "1"
+                             |         | "2"
+    ''')
+    with pytest.raises(InvalidValue) as e:
+        commands.check(context, manifest)
+    assert str(e.value.context['error']) == (
+        "Given enum value 1 does not match property type, "
+        "which is 'integer'."
+    )
+
+
+def test_enum_type_string(tmpdir, rc):
+    context, manifest = load_manifest_and_context(rc, '''
+    d | r | b | m | property | type    | prepare
+    datasets/gov/example     |         |
+                             |         |
+      |   |   | Data         |         |
+      |   |   |   | value    | string  |
+                             | enum    | 1
+                             |         | 2
+    ''')
+    with pytest.raises(InvalidValue) as e:
+        commands.check(context, manifest)
+    assert str(e.value.context['error']) == (
+        "Given enum value 1 does not match property type, "
+        "which is 'string'."
+    )
+
+
+def test_enum_type_none(tmpdir, rc):
+    context, manifest = load_manifest_and_context(rc, '''
+    d | r | b | m | property | type    | source
+    datasets/gov/example     |         |
+                             |         |
+      |   |   | Data         |         |
+      |   |   |   | value    | string  |
+                             | enum    | 1
+                             |         | 2
+    ''')
+    commands.check(context, manifest)